move optional_param
- fixes shared dependency: metrics smoke tests do not want the session fixture, but normal smoke tests also need this helper.
- created tests.helpers.py to solve this
